sign in working, redirects you to profile page
